### 2023-03-20 #5280 @editorialbot
[PRE REVIEW]: rempsyc: Convenience functions for psychology
**Submitting author:** @rempsyc (<a href="http://orcid.org/0000-0003-4315-6788">Rémi Thériault</a>)
**Repository:** https://github.com/rempsyc/rempsyc/
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.1.2
