### 2023-03-21 #5280 @rempsyc
I think anyone from easystats would be well suited to review this paper (in order of relevance: Dominique Makowski, Indrajeet Patil, Daniel Lüdecke, Etienne Bacher, Mattan S. Ben-Shachar, Brenton M. Wiernik). However, note that I am also a member of easystats.

Therefore, alternative relevant reviewers from the list that included both psychology and R in their domains include (starting from bottom of list):
- Psychtek
- brunomontezano
- Metafrancois
- zen-juen
- chartgerink
- katschueler
- trettenbrein
- 62442katieb
- humanfactors
- smasongarrison
- matherion
- raybecker
- avakiai
- navmayan
- navmayan
- navmayan
- ssriv
- ihrke
- mmrabe
- DanielaGawehns
- craddm
- seyhunsaral
- alanfairless
- JonathanReardon
- lindemann09
- janfreyberg
- tomfaulkenberry
- russellpierce
- rmhogervorst
