+++ pkgdown.yml
@@ -12,7 +12,7 @@ articles:
   t-test: t-test.html
   table: table.html
   violin: violin.html
-last_built: 2023-03-22T19:33Z
+last_built: 2023-03-27T15:04Z
 urls:
   reference: https://rempsyc.remi-theriault.com/reference
   article: https://rempsyc.remi-theriault.com/articles
