### 2023-04-03 4d4564e
## rempsyc 0.1.1.3 * `nice_table` and `format_p`: Now support the logical `stars` argument to add significance stars (defaults to `TRUE` for `nice_table` but `FALSE` for `format_p`).
