+++ vignettes/featured.Rmd
@@ -64,3 +64,17 @@ As the `rempsyc` package is starting to gain momentum, an increasing number of e
 # Citations
 
 Jensen, I. N., & Westergaard, M. (2022). Syntax Matters: Exploring the Effect of Linguistic Similarity in Third Language Acquisition. *Language Learning.* https://doi.org/10.1111/lang.12525
+
+# Testimonials
+
+> - I also just want to say how much I love this package and function so much! Thank you so much for your work and contribution! - Seungju
+
+> - Your rempsyc package is amazing! - Ville
+
+> - Thanks for this wonderful and useful package - Maya
+
+> - Your package is amazing! - Hugo
+
+> - I discovered your rempsyc extension to the flextable package and am quite impressed - Tom
+
+> - This package is amazing – thank you so much - Taylor

[…]
