### 2023-04-03 375b611
update flextable required version to CRAN
