+++ DESCRIPTION
@@ -19,7 +19,7 @@ Imports:
     rlang,
     dplyr (>= 1.0.4)
 Suggests: 
-    flextable (>= 0.9.1.007),
+    flextable (>= 0.9.1),
     ggplot2 (>= 3.4.0),
     effectsize (> 0.8.2),
     performance (>= 0.10.0),
@@ -49,7 +49,6 @@ Suggests:
     rmarkdown
 VignetteBuilder: 
     knitr
-Remotes: davidgohel/flextable
 Config/testthat/edition: 3
 Encoding: UTF-8
 LazyData: true
