### 2023-04-17 7172d7c
## rempsyc 0.1.1.4 * Fix a newly introduced bug whereas `nice_normality`, `nice_density`, and `nice_qq` would not work without a group argument.
