+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: rempsyc
 Title: Convenience Functions for Psychology
-Version: 0.1.1.3
-Date: 2023-04-03
+Version: 0.1.1.4
+Date: 2023-04-17
 Authors@R: 
     person("Rémi", "Thériault", , "remi.theriault@mail.mcgill.ca", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-4315-6788"))
