### 2023-04-25 b1be7ab
## rempsyc 0.1.1.4 * `nice_lm` and `nice_lm_slopes`: better error messages if the wrong object type is provided.
