+++ NEWS.md
@@ -1,6 +1,9 @@
 # rempsyc 0.1.x
 * Incoming!
 
+## rempsyc 0.1.1.4
+* `nice_lm` and `nice_lm_slopes`: better error messages if the wrong object type is provided.
+
 ## rempsyc 0.1.1.3
 * `nice_table` and `format_p`: Now support the logical `stars` argument to add significance stars (defaults to `TRUE` for `nice_table` but `FALSE` for `format_p`).
 
