### 2023-04-26 b5c1075
## rempsyc 0.1.1.6
* `nice_contrasts` and `nice_lm_contrasts` gain an argument `boot` (now defaults to FALSE), consistent with `nice_violin`, now allowing not to bootstrap the effect size (thanks to an update in `bootES`)
