+++ README.Rmd
@@ -119,10 +119,18 @@ nice_t_test(data = mtcars,
             response = c("mpg", "disp", "drat", "wt"),
             group = "am") -> t.tests
 t.tests
+```
 
+```{r, results='hide'}
 # Format t-test results
 t_table <- nice_table(t.tests)
 t_table
+```
+
+```{r, include=FALSE}
+table_temp <- flextable::autofit(t_table)
+flextable::save_as_image(table_temp, path = "man/figures/README-nice_t_test-1.png",
+                         expand = 0, res = 300)
 ```
 
 <img src="man/figures/README-nice_t_test-1.png" width="60%" />
@@ -145,9 +153,19 @@ contrasts <- nice_contrasts(
   group = "cyl",
   covariates = "hp")
 contrasts
+```
 
+```{r, results="hide"}
 # Format contrasts results
 nice_table(contrasts, highlight = .001)
+```
+
+```{r, include=FALSE}
+table_temp <- contrasts |>
+  nice_table(highlight = .001) |>
+  flextable::autofit()
+flextable::save_as_image(table_temp, path = "man/figures/README-nice_contrasts-1.png",
+                         expand = 0, res = 300)
 ```
 
 <img src="man/figures/README-nice_contrasts-1.png" width="70%" />
@@ -165,9 +183,19 @@ moderations <- nice_mod(
   predictor = "gear",
   moderator = "wt")
 moderations
+```
 
+```{r, results="hide"}
 # Format moderation results
 nice_table(moderations, highlight = TRUE)
+```
+
+```{r, include=FALSE}
+table_temp <- moderations |>
+  nice_table(highlight = TRUE) |>
+  flextable::autofit()
+flextable::save_as_image(table_temp, path = "man/figures/README-nice_mod-1.png",
+                         expand = 0, res = 300)
 ```
 
 <img src="man/figures/README-nice_mod-1.png" width="70%" />
@@ -179,15 +207,23 @@ Full tutorial: https://rempsyc.remi-theriault.com/articles/moderation
 For more complicated models not supported by `nice_mod`, one can define the model in the traditional way and feed it to `nice_lm` instead. Supports multiple `lm` models as well.
 
 ```{r nice_lm, warning = FALSE, message = FALSE}
-
 model1 <- lm(mpg ~ cyl + wt * hp, mtcars)
 model2 <- lm(qsec ~ disp + drat * carb, mtcars)
-mods <- nice_lm(list(model1, model2))
+mods <- nice_lm(list(model1, model2), standardize = TRUE)
 mods
+```
 
+```{r, results="hide"}
 # Format moderation results
 nice_table(mods, highlight = TRUE)
+```
 
+```{r, include=FALSE}
+table_temp <- mods |>
+  nice_table(highlight = TRUE) |>
+  flextable::autofit()
+flextable::save_as_image(table_temp, path = "man/figures/README-nice_lm-1.png",
+                         expand = 0, res = 300)
 ```
 
 <img src="man/figures/README-nice_lm-1.png" width="70%" />
@@ -205,9 +241,19 @@ simple.slopes <- nice_slopes(
   predictor = "gear",
   moderator = "wt")
 simple.slopes
+```
 
+```{r, results="hide"}
 # Format simple slopes results
 nice_table(simple.slopes)
+```
+
+```{r, include=FALSE}
+table_temp <- simple.slopes |>
+  nice_table() |>
+  flextable::autofit()
+flextable::save_as_image(table_temp, path = "man/figures/README-nice_slopes-1.png",
+                         expand = 0, res = 300)
 ```
 
 <img src="man/figures/README-nice_slopes-1.png" width="80%" />
[…]
