+++ README.md
@@ -118,7 +118,9 @@ t.tests
 #> 2  2.2295592
 #> 3 -1.1245498
 #> 4  2.7329218
+```
 
+``` r
 # Format t-test results
 t_table <- nice_table(t.tests)
 t_table
@@ -157,13 +159,15 @@ contrasts
 #> 5               disp      6 - 8 28 -4.861413 0.000040511099 -3.288726
 #> 6               disp      4 - 6 28 -2.703423 0.011534398020 -1.514296
 #>     CI_lower   CI_upper
-#> 1  2.6313865  4.5520939
-#> 2  0.8069705  1.9501735
-#> 3  1.3395887  3.0175848
-#> 4 -5.7954871 -3.9271356
-#> 5 -4.2824222 -2.1855571
-#> 6 -2.2702868 -0.9346695
+#> 1  2.6216621  4.5307934
+#> 2  0.8116183  2.0311785
+#> 3  1.3501830  3.0545161
+#> 4 -5.7312791 -3.8380868
+#> 5 -4.3194869 -2.2308193
+#> 6 -2.1849423 -0.8902772
+```
 
+``` r
 # Format contrasts results
 nice_table(contrasts, highlight = .001)
 ```
@@ -185,21 +189,23 @@ moderations <- nice_mod(
   predictor = "gear",
   moderator = "wt")
 moderations
-#>   Model Number Dependent Variable Predictor df          b          t          p
-#> 1            1                mpg      gear 28   5.615951  1.9437108 0.06204275
-#> 2            1                mpg        wt 28   1.403861  0.4301493 0.67037970
-#> 3            1                mpg   gear:wt 28  -1.966931 -2.1551077 0.03989970
-#> 4            2               disp      gear 28  35.797623  0.6121820 0.54535707
-#> 5            2               disp        wt 28 160.930043  2.4364098 0.02144867
-#> 6            2               disp   gear:wt 28 -15.037022 -0.8140664 0.42247646
-#>           sr2     CI_lower   CI_upper
-#> 1 0.028488305 0.0000000000 0.08418650
-#> 2 0.001395217 0.0000000000 0.01331121
-#> 3 0.035022025 0.0003502202 0.09723370
-#> 4 0.002737218 0.0000000000 0.01919662
-#> 5 0.043355972 0.0004335597 0.11228977
-#> 6 0.004840251 0.0000000000 0.02679265
+#>   Model Number Dependent Variable Predictor df           B          t
+#> 1            1                mpg      gear 28 -0.08718042 -0.7982999
+#> 2            1                mpg        wt 28 -0.94959988 -8.6037724
+#> 3            1                mpg   gear:wt 28 -0.23559962 -2.1551077
+#> 4            2               disp      gear 28 -0.07488985 -0.6967831
+#> 5            2               disp        wt 28  0.83273987  7.6662883
+#> 6            2               disp   gear:wt 28 -0.08758665 -0.8140664
+#>                   p         sr2     CI_lower   CI_upper
+#> 1 0.431415645312886 0.004805465 0.0000000000 0.02702141
+#> 2 0.000000002383144 0.558188818 0.3142326391 0.80214500
+#> 3 0.039899695159515 0.035022025 0.0003502202 0.09723370
+#> 4 0.491683361920263 0.003546038 0.0000000000 0.02230154
+#> 5 0.000000023731710 0.429258143 0.1916386492 0.66687764
+#> 6 0.422476456495512 0.004840251 0.0000000000 0.02679265
+```
 
+``` r
 # Format moderation results
 nice_table(moderations, highlight = TRUE)
 ```
@@ -215,30 +221,31 @@ the model in the traditional way and feed it to `nice_lm` instead.
 Supports multiple `lm` models as well.
 
 ``` r
-
[…]
