+++ README.Rmd
@@ -133,7 +133,7 @@ flextable::save_as_image(table_temp, path = "man/figures/README-nice_t_test-1.pn
                          expand = 0, res = 300)
 ```
 
-<img src="man/figures/README-nice_t_test-1.png" width="60%" />
+<img src="man/figures/README-nice_t_test-1.png" />
 
 ```{r table_save, warning = FALSE, message = FALSE, eval = FALSE}
 # Save to word
@@ -168,7 +168,7 @@ flextable::save_as_image(table_temp, path = "man/figures/README-nice_contrasts-1
                          expand = 0, res = 300)
 ```
 
-<img src="man/figures/README-nice_contrasts-1.png" width="70%" />
+<img src="man/figures/README-nice_contrasts-1.png" />
 
 Full tutorial: https://rempsyc.remi-theriault.com/articles/contrasts
 
@@ -198,7 +198,7 @@ flextable::save_as_image(table_temp, path = "man/figures/README-nice_mod-1.png",
                          expand = 0, res = 300)
 ```
 
-<img src="man/figures/README-nice_mod-1.png" width="70%" />
+<img src="man/figures/README-nice_mod-1.png" />
 
 Full tutorial: https://rempsyc.remi-theriault.com/articles/moderation
 
@@ -226,7 +226,7 @@ flextable::save_as_image(table_temp, path = "man/figures/README-nice_lm-1.png",
                          expand = 0, res = 300)
 ```
 
-<img src="man/figures/README-nice_lm-1.png" width="70%" />
+<img src="man/figures/README-nice_lm-1.png" />
 
 Full tutorial: https://rempsyc.remi-theriault.com/articles/moderation
 
@@ -256,7 +256,7 @@ flextable::save_as_image(table_temp, path = "man/figures/README-nice_slopes-1.pn
                          expand = 0, res = 300)
 ```
 
-<img src="man/figures/README-nice_slopes-1.png" width="80%" />
+<img src="man/figures/README-nice_slopes-1.png" />
 
 Full tutorial: https://rempsyc.remi-theriault.com/articles/moderation
 
@@ -285,7 +285,7 @@ flextable::save_as_image(table_temp, path = "man/figures/README-nice_lm_slopes-1
                          expand = 0, res = 300)
 ```
 
-<img src="man/figures/README-nice_lm_slopes-1.png" width="80%" />
+<img src="man/figures/README-nice_lm_slopes-1.png" />
 
 Full tutorial: https://rempsyc.remi-theriault.com/articles/moderation
 
