+++ vignettes/featured.Rmd
@@ -67,6 +67,8 @@ Jensen, I. N., & Westergaard, M. (2022). Syntax Matters: Exploring the Effect of
 
 # Testimonials
 
+> - Just discovered your Publication-ready APA tables tutorials. They are Great! Thank you so much - Yayouk
+
 > - I also just want to say how much I love this package and function so much! Thank you so much for your work and contribution! - Seungju
 
 > - Your rempsyc package is amazing! - Ville
