+++ R/nice_violin.R
@@ -10,8 +10,8 @@
 #' For the *easystats* equivalent, see: [see::geom_violindot()].
 #'
 #' @param data The data frame.
-#' @param group The group by which to plot the variable.
 #' @param response The dependent variable to be plotted.
+#' @param group The group by which to plot the variable.
 #' @param boot Logical, whether to use bootstrapping for the confidence
 #' interval or not.
 #' @param bootstraps How many bootstraps to use.
@@ -66,7 +66,6 @@
 #' # Make the basic plot
 #' nice_violin(
 #'   data = ToothGrowth,
-#'   group = "dose",
 #'   response = "len"
 #' )
 #' \donttest{
@@ -191,8 +190,8 @@
 #' @importFrom rlang .data UQ
 
 nice_violin <- function(data,
-                        group,
                         response,
+                        group = NULL,
                         boot = FALSE,
                         bootstraps = 2000,
                         colours,
@@ -224,16 +223,9 @@ nice_violin <- function(data,
     rlang::check_installed(c("boot"), reason = "for this feature.")
   }
 
-  var_message <- "' variable does not seem to exist in this data set... Typo?"
-
-  if (!response %in% names(data)) stop(paste0("The '", response, var_message))
-
-  if (missing(group)) {
-    group <- "All data"
+  if (is.null(group)) {
+    group <- "All"
     data[[group]] <- group
-  } else {
-    if (!group %in% names(data)) stop(paste0("The '", group, var_message))
-    data[[group]] <- as.factor(data[[group]])
   }
 
   data[[response]] <- as.numeric(data[[response]])
