+++ R/nice_normality.R
@@ -77,6 +77,7 @@ nice_normality <- function(data,
   check_col_names(data, c(group, variable))
   rlang::check_installed(c("ggplot2", "see", "patchwork"),
                          reason = "for this function.")
+
   plot.a <- nice_density(
     data = data, variable = variable, group = group,
     colours = colours, groups.labels = groups.labels,
