+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: rempsyc
 Title: Convenience Functions for Psychology
-Version: 0.1.1.5
-Date: 2023-04-25
+Version: 0.1.1.6
+Date: 2023-05-04
 Authors@R: 
     person("Rémi", "Thériault", , "remi.theriault@mail.mcgill.ca", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-4315-6788"))
