### 2023-05-04 9649d62
fix bug with nice_lm_slopes and nice_violin
