+++ vignettes/moderation.Rmd
@@ -174,7 +174,7 @@ nice_slopes(data = mtcars2,
   nice_table(highlight = TRUE)
 ```
 
-2## Complex models: `nice_lm`
+## Complex models: `nice_lm`
 For more complicated models not supported by `nice_mod`, one can define the model in the traditional way and feed it to `nice_lm` and `nice_lm_slopes` instead. They support multiple `lm` models as well.
 
 ### `nice_lm`
