### 2023-05-09 1b6e016
## rempsyc 0.1.1.7 * `format_r`: will now convert NA values to "" instead of "NA"; this is useful when using the `col.format.r` argument in `nice_table()` for correlation matrices.
