+++ NEWS.md
@@ -1,6 +1,9 @@
 # rempsyc 0.1.x
 * Incoming!
 
+## rempsyc 0.1.1.7
+* `format_r`: will now convert NA values to "" instead of "NA"; this is useful when using the `col.format.r` argument in `nice_table()` for correlation matrices.
+
 ## rempsyc 0.1.1.6
 * Fix bug when checking column names (all functions).
 * Fix bug in `nice_violin` to accept NULL group argument again (similar to problem corrected in version 0.1.1.4). Also changed the order of argument, with `response` second, and `group` third, since group becomes optional (as consistent with other similar functions).
