+++ vignettes/featured.Rmd
@@ -33,6 +33,14 @@ As the `rempsyc` package is starting to gain momentum, an increasing number of e
 
 [![](https://rempsyc.remi-theriault.com/images/blog_featured/walker_twitter.png)](https://twitter.com/DrLyndonWalker/status/1595700549927960577)
 
+# Prof. Erik Marsja, PhD
+
+[![](https://rempsyc.remi-theriault.com/images/blog_featured/marsja_twitter.png)](https://twitter.com/MarsjaSe/status/1655460976764805120)
+
+---
+
+[![](https://rempsyc.remi-theriault.com/images/blog_featured/marsja.png)](https://www.marsja.se/report-correlation-in-apa-style-using-r-text-tables/)
+
 # Sharon Machlis
 
 [![](https://rempsyc.remi-theriault.com/images/blog_featured/machlis_title.png)](https://www.infoworld.com/article/3667496/12-ggplot-extensions-for-snazzier-r-graphics.html?page=2)
