### 2023-05-10 e02b3f2
## rempsyc 0.1.1.8 * `nice_table()`: It was requested that the leading zero be omitted for beta values, so that is now the case.
