+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: rempsyc
 Title: Convenience Functions for Psychology
 Version: 0.1.1.7
-Date: 2023-05-09
+Date: 2023-05-10
 Authors@R: 
     person("Rémi", "Thériault", , "remi.theriault@mail.mcgill.ca", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-4315-6788"))
