### 2023-05-11 379f757
## rempsyc 0.1.1.9 * `nice_table`:   * Now capitalize fit indices for a better integration with package {lavaanExtra}.   * Simplification of internal coding and column formatting, e.g., columns names close equivalents, such as `"pvalue"` and `"p.value"`, are now automatically converted to `"p"` (similar for others such as `"chisq"` to `"chi2"`, `conf.low` and `CI_low` to `CI_lower`, etc.), for proper formatting.
