+++ R/nice_table.R
@@ -158,6 +158,33 @@ nice_table <- function(data,
 
   format_p_internal <- ifelse(isTRUE(stars), "format_p_stars", "format_p")
 
+  #   _______________________________________
+  #   Replace possible alternative names ####
+  dataframe <- dataframe %>%
+    rename_with(
+      ~ case_when(
+        . == "p.value" ~ "p",
+        . == "pvalue" ~ "p",
+        . == "chisq" ~ "chi2",
+        . == "conf.low" ~ "CI_lower",
+        . == "conf.high" ~ "CI_upper",
+        . == "CI_low" ~ "CI_lower",
+        . == "CI_high" ~ "CI_upper",
+        . == "df_error" ~ "df",
+        . == "Cohens_d" ~ "d",
+        . == "Coefficient" ~ "b",
+        . == "Std_Coefficient" ~ "B",
+        . == "r_rank_biserial" ~ "rrb",
+        . == "Eta2" ~ "n2",
+        . == "Eta2_partial" ~ "np2",
+        . == "mu" ~ "Mu",
+        . == "term" ~ "Term",
+        . == "method" ~ "Method",
+        . == "alternative" ~ "Alternative",
+        TRUE ~ .
+      )
+    )
+
   #   __________________________________
   #   Broom integration            ####
 
@@ -274,15 +301,9 @@ prepare_broom <- function(dataframe, broom)  {
     dataframe <- dataframe %>%
       rename_with(
         ~ case_when(
-          . == "p.value" ~ "p",
-          . == "conf.low" ~ "CI_lower",
-          . == "conf.high" ~ "CI_upper",
           . == "statistic" ~ "t",
           . == "std.error" ~ "SE",
           . == "parameter" ~ "df",
-          . == "term" ~ "Term",
-          . == "method" ~ "Method",
-          . == "alternative" ~ "Alternative",
           TRUE ~ .
         )
       )
@@ -352,27 +373,9 @@ prepare_report <- function(dataframe, report, short)  {
       }
     }
     dataframe <- dataframe %>%
-      rename_with(
-        ~ case_when(
-          . == "CI_low" ~ "CI_lower",
-          . == "CI_high" ~ "CI_upper",
-          . == "df_error" ~ "df",
-          TRUE ~ .
-        )
-      ) %>%
-      relocate(any_of(c("Method", "Alternative"))) %>%
+      relocate(any_of(c("method", "alternative"))) %>%
       select(-any_of("CI"))
     if (report == "t.test") {
-      dataframe <- dataframe %>%
-        rename_with(
-          ~ case_when(
-            . == "Cohens_d_CI_low" ~ "d_CI_low",
-            . == "Cohens_d_CI_high" ~ "d_CI_high",
-            . == "Cohens_d" ~ "d",
-            . == "mu" ~ "Mu",
-            TRUE ~ .
-          )
-        )
       dataframe <- dataframe %>%
         format_CI(col.name = "95% CI (t)") %>%
         relocate(`95% CI (t)`, .after = t)
@@ -384,19 +387,12 @@ prepare_report <- function(dataframe, report, short)  {
       if (short == TRUE) {
         dataframe <- dataframe %>%
           select_if(!names(.) %in% c(
-            "Method", "Alternative", "Mean_Group1",
+            "method", "alternative", "Mean_Group1",
             "Mean_Group2", "Difference", "95% CI (t)"
           ))
       }
     } else if (report == "lm") {
       dataframe <- dataframe %>%
-        rename_with(
-          ~ case_when(
-            . == "Coefficient" ~ "b",
-            . == "Std_Coefficient" ~ "B",
[…]
