+++ vignettes/t-test.Rmd
@@ -153,7 +153,6 @@ nice_table(stats.table, broom = "t.test")
 library(report)
 (stats.table <- as.data.frame(report(model)))
 nice_table(stats.table, report = "t.test")
-
 ```
 
 The `report` package provides quite comprehensive tables, so one may request an abbreviated table with the `short` argument.
