+++ R/nice_table.R
@@ -679,7 +679,7 @@ format_columns <- function(dataframe, table, italics, separate.header,
       "95% CI (b)", "95% CI (B)", "95% CI (t)", "95% CI (d)",
       "95% CI (np2)", "95% CI (n2)", "95% CI (rrb)", "np2",
       "n2", "ges", "dR", "Predictor (+/-1 SD)", "M1 - M2", "tau",
-      "rho", "rrb", "chi2", "chi2.df"
+      "rho", "rrb", "chi2", "chi2.df", "B"
     ),
     value = c(
       '"95% CI (", flextable::as_i("b"), ")"',
@@ -699,7 +699,8 @@ format_columns <- function(dataframe, table, italics, separate.header,
       '"\u03C1"',
       'flextable::as_i("r"), flextable::as_i(flextable::as_sub("rb"))',
       '"\u03C7", flextable::as_sup("2")',
-      '"\u03C7", flextable::as_sup("2"), "\u2215", flextable::as_i("df")'
+      '"\u03C7", flextable::as_sup("2"), "\u2215", flextable::as_i("df")',
+      '"\u03B2"' # This is beta
     )
   )
   for (i in seq(nrow(compose.table1))) {
@@ -713,10 +714,10 @@ format_columns <- function(dataframe, table, italics, separate.header,
   }
   # Values that can't be greater than 1, but not just italic formatting like r
   compose.table2 <- data.frame(
-    col = c("R2", "sr2", "CFI", "TLI", "RMSEA", "SRMR", "B"),
+    col = c("R2", "sr2", "CFI", "TLI", "RMSEA", "SRMR"),
     value = c('flextable::as_i("R"), flextable::as_sup("2")',
               'flextable::as_i("sr"), flextable::as_sup("2")',
-              "CFI", "TLI", "RMSEA", "SRMR", '"\u03B2"')
+              "CFI", "TLI", "RMSEA", "SRMR")
   )
 
   if(!missing(separate.header)) {
