+++ man/nice_slopes.Rd
@@ -75,7 +75,7 @@ the interpretation of the confidence interval. In \code{rempsyc}, instead
 of using the default one-sided alternative ("greater"), we use the
 two-sided alternative.
 
-To interpret the sr2, use \code{\link[effectsize:interpret_omega_squared]{effectsize::interpret_r2_semipartial()}}.
+To interpret the sr2, use \code{\link[effectsize:interpret_r2_semipartial]{effectsize::interpret_r2_semipartial()}}.
 
 For the \emph{easystats} equivalent, use \code{\link[report:report]{report::report()}} on the \code{\link[=lm]{lm()}}
 model object.
