+++ R/nice_mod.R
@@ -70,7 +70,7 @@
 #' x
 #' @examplesIf requireNamespace("effectsize", quietly = TRUE) & packageVersion("effectsize") >= "0.8.3.5"
 #' # Get interpretations
-#' cbind(x, Interpretation = effectsize::interpret_r2_semipartial(x$sr2))
+#' cbind(x, Interpretation = effectsize::interpret_omega_squared(x$sr2))
 #'
 #' @seealso
 #' Checking simple slopes after testing for moderation:
