### 2023-05-20 9205c2b
correct some ggplot2 warnings about deprecation
