+++ R/nice_density.R
@@ -128,8 +128,8 @@ nice_density <- function(data,
   }
 
   # Make plot
-  plot <- ggplot2::ggplot(data, ggplot2::aes_string(x = variable,
-                                                    fill = group)) +
+  plot <- ggplot2::ggplot(data, ggplot2::aes(x = .data[[variable]],
+                                             fill = .data[[group]])) +
     {
       if (isTRUE(histogram)) {
         ggplot2::geom_histogram(ggplot2::aes(y = ggplot2::after_stat(
@@ -138,13 +138,13 @@ nice_density <- function(data,
         )
       }
     } +
-    ggplot2::geom_density(alpha = 0.6, size = 1, colour = "gray25") +
+    ggplot2::geom_density(alpha = 0.6, linewidth = 1, colour = "gray25") +
     ggplot2::theme_bw(base_size = 24) +
     ggplot2::ggtitle(title) +
     ggplot2::facet_grid(gform) +
     ggplot2::geom_line(
       data = dat_norm, ggplot2::aes(x = x, y = y),
-      color = "darkslateblue", size = 1.2, alpha = 0.9
+      color = "darkslateblue", linewidth = 1.2, alpha = 0.9
     ) +
     ggplot2::ylab(ytitle) +
     ggplot2::xlab(xtitle) +
@@ -175,7 +175,7 @@ nice_density <- function(data,
       if (grid == TRUE) {
         ggplot2::theme(
           panel.grid.major = ggplot2::element_line(),
-          panel.grid.minor = ggplot2::element_line(size = 0.5)
+          panel.grid.minor = ggplot2::element_line(linewidth = 0.5)
         )
       }
     }
