### 2023-05-20 6263e3f
correct beta rounding [skip ci]
