+++ DESCRIPTION
@@ -17,7 +17,7 @@ Depends:
     R (>= 3.6)
 Imports: 
     rlang,
-    dplyr (>= 1.0.4)
+    dplyr (>= 1.1.0)
 Suggests: 
     flextable (>= 0.9.1),
     ggplot2 (>= 3.4.0),
