+++ TODOS.md
@@ -1,5 +1,7 @@
 # Development objectives (to-do list)
 
+-   Make all tests and examples conditional on suggested dependencies availability...
+-   When effectsize new version is on CRAN, update all occurrences of `effectsize::interpret_omega_squared()` to `effectsize::interpret_r2_semipartial()`
 -   `nice_lm_contrasts`: improve function and add to contrasts vignette!
 -   `rcompanion_groupwiseMean`
     -   Rewrite with only the `boot` package and `dplyr::group_by`?
