### 2023-05-23 bbdf556
run vignettes and tests conditionnally
