+++ vignettes/circles.Rmd
@@ -14,9 +14,22 @@ vignette: >
 ---
 
 ```{r, include=FALSE}
+library(knitr)
+
 knitr::opts_chunk$set(fig.width = 6, fig.height = 6, warning = FALSE, 
                       message = FALSE, results = 'hide', out.width = "30%")
 knitr::opts_knit$set(root.dir = tempdir())
+
+pkgs <- c("VennDiagram")
+successfully_loaded <- vapply(pkgs, requireNamespace, FUN.VALUE = logical(1L), quietly = TRUE)
+can_evaluate <- all(successfully_loaded)
+
+if (can_evaluate) {
+  knitr::opts_chunk$set(eval = TRUE)
+  vapply(pkgs, require, FUN.VALUE = logical(1L), quietly = TRUE, character.only = TRUE)
+} else {
+  knitr::opts_chunk$set(eval = FALSE)
+}
 ```
 
 For anyone in social psychology using the Inclusion of the Other in the Self (IOS) Scale to measure self-other merging, you might have wished that you could know for sure how much self-other overlap there is between your participant and a target group or individual.
