+++ vignettes/contrasts.Rmd
@@ -14,9 +14,23 @@ vignette: >
 ---
 
 ```{r global_options, include=FALSE}
+library(knitr)
+
 knitr::opts_chunk$set(fig.width = 7, fig.height = 7, warning = FALSE, 
                       message = FALSE, out.width = "70%")
 knitr::opts_knit$set(root.dir = tempdir())
+
+pkgs <- c("bootES", "emmeans", "flextable", "ggplot2", "boot", "ggsignif",
+          "ggpubr")
+successfully_loaded <- vapply(pkgs, requireNamespace, FUN.VALUE = logical(1L), quietly = TRUE)
+can_evaluate <- all(successfully_loaded)
+
+if (can_evaluate) {
+  knitr::opts_chunk$set(eval = TRUE)
+  vapply(pkgs, require, FUN.VALUE = logical(1L), quietly = TRUE, character.only = TRUE)
+} else {
+  knitr::opts_chunk$set(eval = FALSE)
+}
 ```
 
 ## Basic Idea
