### 2023-05-23 f3e355e
make all examples conditional
