+++ R/nice_lm_slopes.R
@@ -36,7 +36,7 @@
 #' t-value, p-value, and the effect size, the semi-partial correlation
 #' squared, and its confidence interval.
 #' @export
-#' @examples
+#' @examplesIf requireNamespace("effectsize", quietly = TRUE)
 #' # Make and format model
 #' model <- lm(mpg ~ gear * wt, mtcars)
 #' nice_lm_slopes(model, predictor = "gear", moderator = "wt")
