+++ man/nice_lm_contrasts.Rd
@@ -57,6 +57,7 @@ For the \emph{easystats} equivalent, see:
 \code{\link[modelbased:estimate_contrasts]{modelbased::estimate_contrasts()}}.
 }
 \examples{
+\dontshow{if (requireNamespace("bootES", quietly = TRUE) && requireNamespace("emmeans", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 # Make and format model (group need to be a factor)
 model <- lm(mpg ~ as.factor(cyl) + wt * hp, mtcars)
 nice_lm_contrasts(model, group = "cyl", data = mtcars, bootstraps = 500)
@@ -65,7 +66,7 @@ model2 <- lm(qsec ~ as.factor(cyl), data = mtcars)
 my.models <- list(model, model2)
 
 nice_lm_contrasts(my.models, group = "cyl", data = mtcars, bootstraps = 500)
-
+\dontshow{\}) # examplesIf}
 }
 \seealso{
 \code{\link{nice_contrasts}},
