### 2023-05-23 318bc30
more corrections to conditional examples
