+++ R/nice_lm_slopes.R
@@ -46,7 +46,7 @@
 #' my.models <- list(model, model2)
 #' x <- nice_lm_slopes(my.models, predictor = "gear", moderator = "wt")
 #' x
-#' @examplesIf requireNamespace("effectsize", quietly = TRUE) & packageVersion("effectsize") >= "0.8.3.5"
+#' @examplesIf requireNamespace("effectsize", quietly = TRUE) && packageVersion("effectsize") >= "0.8.3.5"
 #' # Get interpretations
 #' cbind(x, Interpretation = effectsize::interpret_omega_squared(x$sr2))
 #'
