+++ R/nice_slopes.R
@@ -72,7 +72,7 @@
 #'   moderator2 = "am"
 #' )
 #' x
-#' @examplesIf requireNamespace("effectsize", quietly = TRUE) & packageVersion("effectsize") >= "0.8.3.5"
+#' @examplesIf requireNamespace("effectsize", quietly = TRUE) && packageVersion("effectsize") >= "0.8.3.5"
 #' # Get interpretations
 #' cbind(x, Interpretation = effectsize::interpret_omega_squared(x$sr2))
 #'
