+++ man/nice_lm_slopes.Rd
@@ -83,7 +83,7 @@ my.models <- list(model, model2)
 x <- nice_lm_slopes(my.models, predictor = "gear", moderator = "wt")
 x
 \dontshow{\}) # examplesIf}
-\dontshow{if (requireNamespace("effectsize", quietly = TRUE) & packageVersion("effectsize") >= "0.8.3.5") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+\dontshow{if (requireNamespace("effectsize", quietly = TRUE) && packageVersion("effectsize") >= "0.8.3.5") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 # Get interpretations
 cbind(x, Interpretation = effectsize::interpret_omega_squared(x$sr2))
 \dontshow{\}) # examplesIf}
