### 2023-06-03 7179274
minor correction to handling of missing values in order.factor in nice_violin
