### 2023-06-05 840b210
# rempsyc 0.1.2.3 * `nice_scatter`: fix bug whereas the group legend was lost even when using `has.legend = TRUE`.
+ styler
