+++ R/nice_assumptions.R
@@ -61,12 +61,15 @@ nice_assumptions <- function(model) {
   names(df) <- c("Model", "shapiro", "bp", "dw")
   df <- df %>%
     dplyr::mutate(dplyr::across(where(is.numeric), round, 3),
-           Diagnostic = rowSums(dplyr::select(., shapiro:dw) < .05))
+      Diagnostic = rowSums(dplyr::select(., shapiro:dw) < .05)
+    )
 
-  names(df) <- c("Model", "Normality (Shapiro-Wilk)",
-                 "Homoscedasticity (Breusch-Pagan)",
-                 "Autocorrelation of residuals (Durbin-Watson)",
-                 "Diagnostic")
+  names(df) <- c(
+    "Model", "Normality (Shapiro-Wilk)",
+    "Homoscedasticity (Breusch-Pagan)",
+    "Autocorrelation of residuals (Durbin-Watson)",
+    "Diagnostic"
+  )
 
   row.names(df) <- NULL
   class(df) <- c("nice_assumptions", class(df))
