+++ README.Rmd
@@ -397,6 +397,10 @@ cormatrix_excel(
 )
 ```
 
+```{r cormatrix_excel_remove, include = FALSE}
+unlink("cormatrix1.xlsx")
+```
+
 <img src="man/figures/cormatrix.png" width="80%" />
 <img src="man/figures/cormatrix2.png" width="80%" />
 
