### 2023-06-05 3d487bd
* `has.legend` now defaults to `TRUE` when using the `group` argument.
