+++ README.Rmd
@@ -355,14 +355,13 @@ nice_scatter(
 )
 ```
 
-```{r nice_scatter2, warning = FALSE, message = FALSE, fig.width=10, fig.height=7, out.width="60%"}
+```{r nice_scatter2, warning = FALSE, message = FALSE, fig.width=10, fig.height=7, out.width="70%"}
 nice_scatter(
   data = mtcars,
   predictor = "wt",
   response = "mpg",
   group = "cyl",
-  has.confband = TRUE,
-  has.legend = TRUE
+  has.confband = TRUE
 )
 ```
 
