### 2023-06-05 38dbfae
remove legend in vignette [skip ci]
