+++ vignettes/scatter.Rmd
@@ -212,18 +212,6 @@ nice_scatter(
 )
 ```
 
-### Add a legend
-
-```{r}
-nice_scatter(
-  data = mtcars,
-  predictor = "wt",
-  response = "mpg",
-  group = "cyl",
-  has.legend = TRUE
-)
-```
-
 ### Change order of labels on the legend
 
 ```{r}
