### 2023-06-05 a2e57b0
remove all legend calls in vignette [skip ci]
