+++ vignettes/scatter.Rmd
@@ -220,7 +220,6 @@ nice_scatter(
   predictor = "wt",
   response = "mpg",
   group = "cyl",
-  has.legend = TRUE,
   groups.order = c(8, 4, 6)
 )
 # These are the levels of 'mtcars$cyl', so we place lvl 8
@@ -235,7 +234,6 @@ nice_scatter(
   predictor = "wt",
   response = "mpg",
   group = "cyl",
-  has.legend = TRUE,
   groups.labels = c("Weak", "Average", "Powerful")
 )
 # Warning: This applies after changing order of level
@@ -255,7 +253,6 @@ nice_scatter(
   predictor = "wt",
   response = "mpg",
   group = "cyl",
-  has.legend = TRUE,
   legend.title = "Cylinders"
 )
 ```
@@ -304,7 +301,6 @@ nice_scatter(
   predictor = "wt",
   response = "mpg",
   group = "cyl",
-  has.legend = TRUE,
   legend.title = "Cylinders",
   has.linetype = TRUE,
   has.shape = TRUE,
@@ -342,7 +338,6 @@ nice_scatter(
   r.x = 5.5,
   r.y = 25,
   colours = c("burlywood", "darkgoldenrod", "chocolate"),
-  has.legend = TRUE,
   legend.title = "Cylinders",
   groups.labels = c("Weak", "Average", "Powerful")
 )
@@ -373,7 +368,6 @@ Then we need to create a FIRST layer of just the slopes. We add transparency to 
   predictor = "wt",
   response = "mpg",
   has.points = FALSE,
-  has.legend = TRUE,
   group = "cyl",
   colours = c("black", "#00BA38", "#619CFF", "#F8766D"),
   # We add colours manually because we want average to be black to stand out
