### 2023-06-06 ed50917
* New argument `method`, with choices `"lm"` (default) or `"loess"`.
