### 2023-06-06 b462fc8
add comment to special symbols for nice_table code [skip ci]
