### 2023-06-06 f4fb5f3
nice_table: Set attributes for variables which digits should not be changed for later step
