### 2023-06-06 fd1a421
fix vignette pkg check typo
