### 2023-06-07 rempsyc/rempsyc#15 @kinleyid
Comments on JOSS submission
This is a useful library that is clearly already accelerating many researchers' work and reducing the learning curve for those getting started in R. In my review checklist, there are a few items that I've left unchecked: https://github.com/openjournals/joss-reviews/issues/5466#issuecomment-1555922419

1.
    > Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

   I'm not sure that the stated dependencies on [CRAN](https://cran.r-project.org/web/packages/rempsyc/index.html) are correct. I tried to run a few functions (`nice_table` and `nice_violin`) and got various errors that were resolved by updating dplyr from 1.0.8 to 1.1.0, whereas the stated dependency is >1.0.4.

2.
    > Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

   It seems that only the `nice_violin` function is tested in tests/testthat/. Would it not be better to test all of the functions?

3.
    > Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

   I was unable to find any explicit guidelines for those looking to contribute---I suppose it's obvious that anyone can open a pull request, but it might be worth explicitly stating this somewhere.

Additionally, I have a few minor comments:

* The suggested workflow for tables, using `flextable::save_as_docx`, seems a little bit more cumbersome than the method suggested [here](https://github.com/davidgohel/flextable/issues/117#issuecomment-487126438). `print(my_table, preview ="docx")` opens the flextable in a temporary Word document from which it can be quickly copied and pasted into a manuscript. Perhaps both methods could be mentioned.
* The interpolation of the Inclusion of Other in the Self (IOS) scale points should be explained a little bit more. Specifically, does overlap mean linear overlap (which would scale linearly with the distance between the two circles' centres), or is it a measure of the actual overlapping area of the two circles (which would be measured like [this](https://mathworld.wolfram.com/Circle-CircleIntersection.html))?
