### 2023-06-17 3374429
correct bug in grouping for nice_scatter
