### 2023-06-19 2f22ae3
# rempsyc 0.1.2.5 * Fixed bug with `check_installed` where required package version were not being specified correctly. * A way to install all dependencies at once are now provided in the README file. * Contribution Guidelines are now available.
