+++ R/nice_lm_slopes.R
@@ -65,7 +65,9 @@ nice_lm_slopes <- function(model,
                            mod.id = TRUE,
                            ci.alternative = "two.sided",
                            ...) {
-  rlang::check_installed("effectsize", reason = "for this function.")
+  rlang::check_installed("effectsize",
+                         version = "0.8.2",
+                         reason = "for this function.")
   if (inherits(model, "list") && all(unlist(lapply(model, inherits, "lm")))) {
     models.list <- model
   } else if (inherits(model, "lm")) {
