+++ R/nice_slopes.R
@@ -96,7 +96,9 @@ nice_slopes <- function(data,
                         ci.alternative = "two.sided",
                         ...) {
   check_col_names(data, c(predictor, response, moderator, moderator2, covariates))
-  rlang::check_installed("effectsize", reason = "for this function.")
+  rlang::check_installed("effectsize",
+                         version = "0.8.2",
+                         reason = "for this function.")
 
   if (!missing(b.label)) {
     message(paste(
