+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: rempsyc
 Title: Convenience Functions for Psychology
-Version: 0.1.2.5
+Version: 0.1.3
 Date: 2023-06-19
 Authors@R: 
     person("Rémi", "Thériault", , "remi.theriault@mail.mcgill.ca", role = c("aut", "cre"),
