+++ R/nice_lm_slopes.R
@@ -66,7 +66,7 @@ nice_lm_slopes <- function(model,
                            ci.alternative = "two.sided",
                            ...) {
   rlang::check_installed("effectsize",
-                         version = "0.8.2",
+                         version = get_dep_version("effectsize"),
                          reason = "for this function.")
   if (inherits(model, "list") && all(unlist(lapply(model, inherits, "lm")))) {
     models.list <- model
