+++ R/plot_outliers.R
@@ -63,7 +63,7 @@ plot_outliers <- function(data,
                           ...) {
   check_col_names(data, c(group, response))
   rlang::check_installed(c("ggplot2"),
-                         version = "3.4.0",
+                         version = get_dep_version("ggplot2"),
                          reason = "for this function.")
   mtd <- switch(method,
     "mad" = "median",
