+++ R/nice_violin.R
@@ -233,7 +233,7 @@ nice_violin <- function(data,
                         xlabels.angle = 0) {
   check_col_names(data, c(group, response))
   rlang::check_installed(c("ggplot2"),
-                         version = "3.4.0",
+                         version = get_dep_version("ggplot2"),
                          reason = "for this function.")
   if (isTRUE(boot)) {
     rlang::check_installed(c("boot"), reason = "for this feature.")
