### 2023-06-20 #5466 @samhforbes
Thanks @avakiai and @kinleyid for raising these issues - @rempsyc feel free to address those and liaise with the reviewers in the repo, or link back here if you feel anything is unclear.
@brunomontezano it looks like you've also got some issues to raise based on unticked boxes, so feel free to raise issue(s) in the software repo, or ask here if it's bigger picture.
