### 2023-06-20 d53a999
correct get_dep_version
