+++ R/utils.R
@@ -56,10 +56,14 @@ model_is_standardized <- function(models.list) {
 }
 
 #' @noRd
-get_dep_version <- function(dep) {
-  suggests.field <- utils::packageDescription("rempsyc", fields = "Suggests")
-  suggests.list <- strsplit(suggests.field, ",\n")[[1]]
+get_dep_version <- function(dep, pkg = "rempsyc") {
+  suggests.field <- utils::packageDescription(pkg, fields = "Suggests")
+  suggests.list <- unlist(strsplit(suggests.field, ",", fixed = TRUE))
   dep.string <- grep(dep, suggests.list, value = TRUE)
   dep.string <- sub(".*\\((.*?)\\).*", "\\1", dep.string)
-  gsub("[^0-9.]+", "", dep.string)
+  out <- gsub("[^0-9.]+", "", dep.string)
+  if (out == "") {
+    out <- NULL
+  }
+  out
 }
