### 2023-06-22 e30ee4f
update overlap_circle description
