+++ vignettes/circles.Rmd
@@ -34,19 +34,19 @@ if (can_evaluate) {
 }
 ```
 
-For anyone in social psychology using the Inclusion of the Other in the Self (IOS) Scale to measure self-other merging, you might have wished that you could know for sure how much self-other overlap there is between your participant and a target group or individual.
+For anyone in social psychology using the Inclusion of the Other in the Self (IOS) Scale to measure self-other merging, you might have wished that you could know for sure how much actual percentage self-other overlap there is between your participant and a target group or individual.
 
 Indeed, the IOS response choices go from 1 to 7 only, and unfortunately these don't match the percentage overlap from the circle images (i.e., 1/7 is not 14% overlap, and 7/7 is not 100% overlap). You can observe this on the original IOS scale below.
 
 ![](https://rempsyc.remi-theriault.com/images/blog_circles/IOS.png)
 
-So through trial and error (using the `VennDiagram` package), I was able to determine the approximate actual overlap from the pictures (below).
+So through trial and error (using the `VennDiagram` package), I was able to determine the approximate actual *area* overlap from the pictures (below).
 
 ![](https://rempsyc.remi-theriault.com/images//blog_circles/overlap.png)
 
-I feel like the overlap is pretty close to the original scale. Based on this, I used the `approx()` function to interpolate any responses so that 1 = 0% (overlap), 2 = 10%, 3 = 20%, 4 = 30%, 5 = 55%, 6 = 65%, and 7 (the maximum) = 85%.
+I feel like the area overlap is pretty close to the original scale. Based on this, I used the `approx()` function to interpolate any responses so that 1 = 0% (overlap), 2 = 10%, 3 = 20%, 4 = 30%, 5 = 55%, 6 = 65%, and 7 (the maximum) = 85%.
 
-But that also means that a continuous response (for example based on a group average) can also be plotted accurately (e.g., a score of 6.84 would turn into 81.8 % overlap). Let's see a few examples.
+But that also means that a continuous response (for example based on a group average) can also be plotted accurately (e.g., a score of 6.84 would turn into 81.8% area overlap). Let's see a few examples.
 
 
 ### Getting started
@@ -113,5 +113,5 @@ This allowed me for instance, for one study (https://doi.org/10.1177/17470218211
 ![](https://rempsyc.remi-theriault.com/images//blog_circles/figure.png)
 
 ### Thanks for checking in
-    
+
 Make sure to check out this page again if you use the code after a time or if you encounter errors, as I periodically update or improve the code. Feel free to contact me for comments, questions, or requests to improve this function at [https://github.com/rempsyc/rempsyc/issues](https://github.com/rempsyc/rempsyc/issues). See all tutorials here: [https://remi-theriault.com/tutorials](https://remi-theriault.com/tutorials).
