### 2023-06-22 e89a1bf
# rempsyc 0.1.3.1 * Added tests for all functions * Accordingly, updated code to stop using deprecated functions from other packages (`dplyr` and `ggplot2`).
