+++ tests/testthat/test-winsorize_mad.R
@@ -1,0 +1,3 @@
+test_that("winsorize_mad", {
+  expect_snapshot(winsorize_mad(mtcars$qsec, criteria = 2))
+})
