+++ tests/testthat/test-nice_scatter.R
@@ -1,0 +1,283 @@
+test_that("nice_scatter", {
+  skip_if_not_installed("ggplot2")
+
+  # Make the basic plot
+  x1 <- nice_scatter(
+    data = mtcars,
+    predictor = "wt",
+    response = "mpg"
+  )
+
+  expect_s3_class(
+    x1,
+    c("gg", "ggplot2")
+  )
+
+  ggplot2::ggsave("plot.jpg",
+                  width = 7, height = 7, unit = "in",
+                  dpi = 300, path = NULL
+  )
+
+  expect_snapshot_file("plot.jpg")
+
+  # Remove file
+  unlink("plot.jpg")
+
+  # Change x- and y- axis labels
+  x2 <- nice_scatter(
+    data = mtcars,
+    predictor = "wt",
+    response = "mpg",
+    ytitle = "Miles/(US) gallon",
+    xtitle = "Weight (1000 lbs)"
+  )
+
+  expect_s3_class(
+    x2,
+    c("gg", "ggplot2")
+  )
+
+  # Have points "jittered", loess method
+  x3 <- nice_scatter(
+    data = mtcars,
+    predictor = "wt",
+    response = "mpg",
+    has.jitter = TRUE,
+    method = "loess"
+  )
+
+  expect_s3_class(
+    x3,
+    c("gg", "ggplot2")
+  )
+
+  # Change the transparency of the points
+  x4 <- nice_scatter(
+    data = mtcars,
+    predictor = "wt",
+    response = "mpg",
+    alpha = 1
+  )
+
+  expect_s3_class(
+    x4,
+    c("gg", "ggplot2")
+  )
+
+  # Remove points
+  x5 <- nice_scatter(
+    data = mtcars,
+    predictor = "wt",
+    response = "mpg",
+    has.points = FALSE,
+    has.jitter = FALSE
+  )
+
+  expect_s3_class(
+    x5,
+    c("gg", "ggplot2")
+  )
+
+  # Add confidence band
+  x6 <- nice_scatter(
+    data = mtcars,
+    predictor = "wt",
+    response = "mpg",
+    has.confband = TRUE
+  )
+
+  expect_s3_class(
+    x6,
+    c("gg", "ggplot2")
+  )
+
+  # Set x- and y- scales manually
+  x7 <- nice_scatter(
+    data = mtcars,
+    predictor = "wt",
+    response = "mpg",
+    xmin = 1,
+    xmax = 6,
+    xby = 1,
+    ymin = 10,
+    ymax = 35,
+    yby = 5
+  )
+
+  expect_s3_class(
+    x7,
+    c("gg", "ggplot2")
+  )
+
+  # Change plot colour
+  x8 <- nice_scatter(
+    data = mtcars,
+    predictor = "wt",
+    response = "mpg",
+    colours = "blueviolet"
+  )
+
+  expect_s3_class(
+    x8,
+    c("gg", "ggplot2")
+  )
+
+  # Add correlation coefficient to plot and p-value
+  x9 <- nice_scatter(
+    data = mtcars,
+    predictor = "wt",
+    response = "mpg",
+    has.r = TRUE,
+    has.p = TRUE
+  )
+
+  expect_s3_class(
+    x9,
+    c("gg", "ggplot2")
+  )
+
+  # Change location of correlation coefficient or p-value
+  x10 <- nice_scatter(
+    data = mtcars,
+    predictor = "wt",
+    response = "mpg",
+    has.r = TRUE,
+    r.x = 4,
+    r.y = 25,
+    has.p = TRUE,
+    p.x = 5,
+    p.y = 20
+  )
+
+  expect_s3_class(
+    x10,
+    c("gg", "ggplot2")
+  )
+
+  # Plot by group
+  x11 <- nice_scatter(
[…]
