+++ R/nice_qq.R
@@ -119,7 +119,7 @@ nice_qq <- function(data,
     if (grid == TRUE) {
       ggplot2::theme(
         panel.grid.major = ggplot2::element_line(),
-        panel.grid.minor = ggplot2::element_line(size = 0.5)
+        panel.grid.minor = ggplot2::element_line(linewidth = 0.5)
       )
     }
   }

[…]
