### 2023-06-22 rempsyc/rempsyc#15 @rempsyc
Dear Isaac,

First, a sincere thank you for this very helpful and constructive review. You will find my answers below:

> I'm not sure that the stated dependencies on [CRAN](https://cran.r-project.org/web/packages/rempsyc/index.html) are correct. I tried to run a few functions (nice_table and nice_violin) and got various errors that were resolved by updating dplyr from 1.0.8 to 1.1.0, whereas the stated dependency is >1.0.4.

Thanks, this has been resolved in #14 (as of development version 0.1.2.1, and CRAN version 0.1.3).

> It seems that only the nice_violin function is tested in tests/testthat/. Would it not be better to test all of the functions?

Thanks, I have now added tests for all functions (version 0.1.3.1).

> I was unable to find any explicit guidelines for those looking to contribute---I suppose it's obvious that anyone can open a pull request, but it might be worth explicitly stating this somewhere.

Thanks, this has now been added (accessible through the README page): https://rempsyc.remi-theriault.com/CONTRIBUTING.html

> The interpolation of the Inclusion of Other in the Self (IOS) scale points should be explained a little bit more. Specifically, does overlap mean linear overlap (which would scale linearly with the distance between the two circles' centres), or is it a measure of the actual overlapping area of the two circles (which would be measured like [this](https://mathworld.wolfram.com/Circle-CircleIntersection.html))?

Thank you, I have now clarified this point. In the function description:

```r
#' @description Interpolating the Inclusion of the Other in
#' the Self Scale (IOS; self-other merging) easily. The user provides
#' the IOS score, from 1 to 7, and the function will provide a
#' percentage of actual area overlap between the two circles
#' (i.e., not linear overlap), so it is possible to say, e.g.,
#' that experimental group 1 had an average overlap of X%
#' with the other person, whereas experimental group 2 had
#' an average overlap of X% with the other person.
#'
#' @details
#' The circles are generated through the
#' `VennDiagram::draw.pairwise.venn()` function and the desired
#' percentage overlap is passed to its `cross.area` argument
#' ("The size of the intersection between the sets"). The percentage
#' overlap values are interpolated from this reference grid:
#' Score of 1 = 0%, 2 = 10%, 3 = 20%, 4 = 30%, 5 = 55%, 6 = 65%,
#' 7 = 85%.
```

In the vignette and paper, I have also added the keyword "overlap _area_" at a few places to avoid ambiguity. Do you feel like this is enough, or would you add something?

By the way, GREAT work on your jsPych plugin for a continuous IOS. Would you be OK if I added a resource link to your plugin in the documentation `overlap_circle()`? If so, would you prefer that I share your [home page](https://kinleyid.github.io/) or [github project](https://github.com/jspsych/jspsych-contrib/tree/main/packages/plugin-ios) URL?

---

Finally, would you have any comments or suggestions on the following points:

- Content, structure, or organization of the paper
- Names of the functions (based on conventions, etc.)
- Names of the arguments within the functions (based on conventions, etc.)

Thanks again.
