### 2023-06-22 b5628ed
remove expect_snapshot_file, as it creates problems in tests
