+++ R/nice_table.R
@@ -4,6 +4,10 @@
 #' around the `flextable` package with sensical defaults and
 #' automatic formatting features.
 #'
+#' @details The resulting `flextable` objects can be opened in
+#' Word with `print(table, preview ="docx")`, or saved to
+#' Word with the `flextable::save_as_docx()` function.
+#'
 #' @param data The data frame, to be converted to a flextable.
 #' The data frame cannot have duplicate column names.
 #' @param italics Which columns headers should be italic? Useful
