+++ tests/testthat/test-cormatrix_excel.R
@@ -1,4 +1,7 @@
 test_that("cormatrix_excel", {
+  skip_if_not_installed("correlation")
+  skip_if_not_installed("openxlsx2")
+
   expect_snapshot(cormatrix_excel(mtcars, select = c("mpg", "cyl", "disp", "hp", "carb"),
                                   filename = "cormatrix1"))
   expect_snapshot(cormatrix_excel(iris, p_adjust = "none", filename = "cormatrix2"))
