### 2023-06-23 bc08017
Add info on simple slopes
