+++ vignettes/moderation.Rmd
@@ -106,6 +106,8 @@ flextable::save_as_docx(my_table, path = "moderations.docx")
 
 ## Simple slopes: `nice_slopes`
 
+You might have heard about "simple slopes" before. But what does that mean? Essentially, this means looking at the strength (regression coefficient) and significance of the slope, when subsetting for observations that are high, low, or average on a variable, typically the moderating variable. A bit further down, this will get clearer by looking at the plot of the interaction, which shows one slope for observations that are high on the wt (moderating) variable, a second slope for those that are low, and a third slope for those that are average.
+
 Let's extract the simple slopes now, including the sr2.
 
 ```{r}
