### 2023-06-29 f24aac6
Handle error "sample size must be between 3 and 5000" in stats::shapiro.test()
