### 2023-06-29 8c9296f
add changes in nice_assumptions to NEWS.md
