+++ NEWS.md
@@ -2,7 +2,7 @@
 * Incoming!
 
 # rempsyc 0.1.3.3
-* sjewo
+* `nice_assumptions`: p-values from `shapiro.test()` are defined only for vectors of length between 3 and 5000. `nice_assumptions' now checks for the sample size and returns `NA` for samples outside this range (#17, @sjewo).
 
 # rempsyc 0.1.3.2
 * New functions:
