### 2023-06-29 faf4cd1
error "sample size must be between 3 and 5000" in stats::shapiro (#17)
* Handle error "sample size must be between 3 and 5000" in stats::shapiro.test()

* add informative message

* Add tests

* add changes in nice_assumptions to NEWS.md

---------

Co-authored-by: rempsyc <remi.theriault@mail.mcgill.ca>
