+++ NEWS.md
@@ -2,7 +2,7 @@
 * Incoming!
 
 # rempsyc 0.1.3.3
-* `nice_assumptions`: p-values from `shapiro.test()` are defined only for vectors of length between 3 and 5000. `nice_assumptions' now checks for the sample size and returns `NA` for samples outside this range (#17, @sjewo).
+* `nice_assumptions`: *p* values from `shapiro.test()` are defined only for vectors of length between 3 and 5000 and was thus throwing an error in these cases. `nice_assumptions` now checks for the sample size and returns `NA` for samples outside this range (#17, @sjewo).
 
 # rempsyc 0.1.3.2
 * New functions:
