+++ R/nice_contrasts.R
@@ -12,10 +12,6 @@
 #' @param ... Arguments passed to [bootES::bootES].
 #' @keywords planned contrasts pairwise comparisons
 #' group differences
-#' @return A dataframe, with the selected dependent variable(s), comparisons of
-#'         interest, degrees of freedom, t-values, p-values, robust Cohen's d
-#'         (dR), and the lower and upper 95% confidence intervals of the
-#'         effect size (i.e., dR).
 #' @export
 #' @examplesIf requireNamespace("bootES", quietly = TRUE) && requireNamespace("emmeans", quietly = TRUE)
 #' # Basic example
