### 2023-07-02 afe19d2
# rempsyc 0.1.3.4 * `nice_lm_contrasts` and `nice_contrasts`: Code simplified and now supports more than 3 groups, through easystats' `modelbased` package.
