+++ R/nice_contrasts.R
@@ -13,8 +13,9 @@
 #' @keywords planned contrasts pairwise comparisons
 #' group differences
 #' @export
-#' @examplesIf requireNamespace("bootES", quietly = TRUE) && requireNamespace("emmeans", quietly = TRUE)
+#' @examplesIf requireNamespace("bootES", quietly = TRUE) && requireNamespace("modelbased", quietly = TRUE)
 #' # Basic example
+#' set.seed(100)
 #' nice_contrasts(
 #'   data = mtcars,
 #'   response = "mpg",
@@ -22,6 +23,7 @@
 #'   bootstraps = 200
 #' )
 #' \donttest{
+#' set.seed(100)
 #' nice_contrasts(
 #'   data = mtcars,
 #'   response = "disp",
@@ -29,6 +31,7 @@
 #' )
 #'
 #' # Multiple dependent variables
+#' set.seed(100)
 #' nice_contrasts(
 #'   data = mtcars,
 #'   response = c("mpg", "disp", "hp"),
@@ -36,12 +39,25 @@
 #' )
 #'
 #' # Adding covariates
+#' set.seed(100)
 #' nice_contrasts(
 #'   data = mtcars,
 #'   response = "mpg",
 #'   group = "cyl",
 #'   covariates = c("disp", "hp")
 #' )
+#'
+#' # Now supports more than 3 levels
+#' mtcars2 <- mtcars
+#' mtcars2$carb <- as.factor(mtcars2$carb)
+#' set.seed(100)
+#' nice_contrasts(
+#'   data = mtcars,
+#'   response = "mpg",
+#'   group = "carb",
+#'   bootstraps = 200
+#' )
+#'
 #' }
 #' @seealso
 #' \code{\link{nice_lm_contrasts}},
@@ -57,7 +73,7 @@ nice_contrasts <- function(response,
                            bootstraps = 2000,
                            ...) {
   check_col_names(data, c(group, response, covariates))
-  rlang::check_installed(c("bootES", "emmeans"), reason = "for this function.")
+  rlang::check_installed(c("bootES", "modelbased"), reason = "for this function.")
   data[[group]] <- as.factor(data[[group]])
   data[response] <- lapply(data[response], as.numeric)
   if (!missing(covariates)) {
@@ -71,86 +87,9 @@ nice_contrasts <- function(response,
     covariates.term
   ) ### Add support for moderator(s)
   models.list <- lapply(formulas, lm, data = data)
-  leastsquare.list <- lapply(models.list, emmeans::emmeans, group, data = data)
-  groups.contrasts <- list(
-    comp1 = stats::setNames(c(1, 0, -1), levels(data[[group]])),
-    # Add support x groups
-    comp2 = stats::setNames(c(0, 1, -1), levels(data[[group]])),
-    # Add support x groups
-    comp3 = stats::setNames(c(1, -1, 0), levels(data[[group]]))
-  )
 
-  # Add support x groups
-  es.lists <- lapply(groups.contrasts, function(y) {
-    lapply(response, function(x) {
-      bootES::bootES(
-        data = stats::na.omit(data[, c(group, x)]),
-        R = bootstraps,
-        data.col = x,
-        group.col = group,
-        contrast = y,
-        effect.type = effect.type,
-        ...
-      )
-    })
-  })
-
-  contrval.list <- lapply(leastsquare.list, emmeans::contrast,
-    groups.contrasts,
-    adjust = "none"
-  )
-  contrval.sums <- lapply(contrval.list, summary)
-
-  boot.sums <- lapply(seq(length(response)), function(y) {
-    lapply(es.lists, function(x) {
-      as.data.frame(summary(x[[y]]))
-    }) %>%
-      bind_rows()
-  })
[…]
