+++ NEWS.md
@@ -1,6 +1,9 @@
 # rempsyc 0.1.4
 * Incoming!
 
+# rempsyc 0.1.3.4
+* `nice_lm_contrasts` and `nice_contrasts`: Code simplified and now supports more than 3 groups, through easystats' `modelbased` package.
+
 # rempsyc 0.1.3.3
 * `nice_assumptions`: *p* values from `shapiro.test()` are defined only for vectors of length between 3 and 5000 and was thus throwing an error in these cases. `nice_assumptions` now checks for the sample size and returns `NA` for samples outside this range (#17, @sjewo).
 

[…]
