+++ R/nice_lm_contrasts.R
@@ -15,7 +15,10 @@
 #' ([source](https://web.pdx.edu/~newsomj/uvclass/ho_planned%20contrasts.pdf)).
 #'
 #' The effect size and confidence interval are calculated via
-#' [bootES::bootES].
+#' [bootES::bootES], and correct for contrasts but not for covariates and
+#' other predictors. Because this method uses bootstrapping, it is recommended
+#' to set a seed before using for reproducibility reasons (e.g.,
+#' `sed.seet(100)`).
 #'
 #' Does not for the moment support nested comparisons for marginal means,
 #' only a comparison of all groups. For nested comparisons, please use
