+++ vignettes/featured.Rmd
@@ -77,6 +77,8 @@ Jensen, I. N., & Westergaard, M. (2022). Syntax Matters: Exploring the Effect of
 
 # Testimonials
 
+> - I have been using your rempsyc package in R to generate tables, to which I am extremely grateful. Thank you for having come up with such useful package - Keith
+
 > - Thank you for providing this useful package! - Sebastian
 
 > - Thank you for creating such a helpful package for APA tables - I hope this will be a life saver for my thesis formatting - Michaela
