### 2023-07-11 606bb87
# rempsyc 0.1.3.5 * `nice_table`: fix bug when providing a table of more than one `report::report` object made from Wilcoxon tests.
