### 2023-07-11 rempsyc/rempsyc#16 @avakiai
Hi Remi,
Apologies for my late reply. This looks great! Thank you for implementing those changes.

I have no further comments - naming conventions and the paper seem ok to me. Vignettes are also well-written.

I will now check those boxes I had left unchecked in my inital review.

Best wishes,
Ava
