### 2023-07-11 62bde9b
which(is.na(dataframe$Parameter))[1]
