### 2023-07-12 rempsyc/rempsyc#15 @kinleyid
Hi Rémi,

This looks great! I think you've addressed all my comments well. I think that using the word "area" is enough to avoid ambiguity---it's a pretty univocal term. I would be happy for you to mention the jsPsych IOS plugin if you think it would be of interest to your users! Probably the link to the github project would make the most sense.

I have no specific comments on the manuscript's writing or on the function/argument names---I think these are all fine as-is. My only general suggestion would be, for wrapper functions to ggplot, to use argument names that match ggplot's as closely as possible. This way users don't have to learn a second set of conventions if they're already somewhat familiar with ggplot.

Cheers,
Isaac
