### 2023-07-12 rempsyc/rempsyc#15 @rempsyc
Thanks a lot! I like the idea of using `ggplot2` original names as arguments. Some universal arguments do translate directly already, like `alpha`, `method`, and `legend.title`, but perhaps `colours` could be changed to `color`, and `border.size` to `linewidth`.

But I'm a bit torn on this. It was an intentional choice when I started the package to give more sensical names for beginners that *don't* know `ggplot2`. This is also my primary audience, as those familiar with `ggplot2` are more likely to use `ggplot2` directly—I think. But I don't know, maybe even `ggplot2` experts could like the convenience of these plotting functions! Thanks again for your input.
