+++ R/overlap_circle.R
@@ -61,6 +61,11 @@
 #' @seealso
 #' Tutorial: \url{https://rempsyc.remi-theriault.com/articles/circles}
 #'
+#' For a javascript web plugin of a continuous version of the Inclusion of
+#' Other in the Self (IOS) task (instead of the pen and paper version), for
+#' experiments during data collection, rather than data analysis, please see:
+#' \url{https://github.com/jspsych/jspsych-contrib/tree/main/packages/plugin-ios}
+#'
 
 overlap_circle <- function(response,
                            categories = c("Self", "Other"),
