+++ R/rcompanion_groupwiseMean.R
@@ -79,7 +79,7 @@
 #'
 #' @export
 #' @author Salvatore Mangiafico, \email{mangiafico@njaes.rutgers.edu}
-#' @references \url{http://rcompanion.org/handbook/C_03.html}
+#' @references \url{https://rcompanion.org/handbook/C_03.html}
 #' @concept mean confidence interval bootstrap
 #' @return A data frame of requested statistics by group.
 #' @keywords group means confidence intervals bootstrapping internal
