+++ NEWS.md
@@ -1,23 +1,23 @@
 # rempsyc 0.1.4
 * CRAN submission
 
-# rempsyc 0.1.3.5
+## rempsyc 0.1.3.5
 * `nice_table`: fix bug when providing a table of more than one `report::report` object made from Wilcoxon tests.
 
-# rempsyc 0.1.3.4
+## rempsyc 0.1.3.4
 * `nice_lm_contrasts` and `nice_contrasts`: Code simplified and now supports more than 3 groups, through easystats' `modelbased` package.
 
-# rempsyc 0.1.3.3
+## rempsyc 0.1.3.3
 * `nice_assumptions`: *p* values from `shapiro.test()` are defined only for vectors of length between 3 and 5000 and was thus throwing an error in these cases. `nice_assumptions` now checks for the sample size and returns `NA` for samples outside this range (#17, @sjewo).
 
-# rempsyc 0.1.3.2
+## rempsyc 0.1.3.2
 * New functions:
   * `install_if_not_installed`: will only install specified packages if they are not installed. 
   * `get_dep_version`: checks the required version of a specific package dependency by looking at the specified package's DESCRIPTION file directly.
 * `best_duplicate`: now supports the `keep.rows` argument, to add a column containing the original rows.
 * `overlap_circle`: now support two other scoring methods: "percentage" (1 to 100), and "direct" (must provide the three values from circle 1, overlapping area, and circle 2).
 
-# rempsyc 0.1.3.1
+## rempsyc 0.1.3.1
 * Added tests for all functions
 * Accordingly, updated code to stop using deprecated functions from other packages (`dplyr` and `ggplot2`).
 
