### 2023-07-20 #5466 @rempsyc
Thanks @samhforbes

- The release for the latest version, `v1.0.1.4`, is available here: https://github.com/rempsyc/rempsyc/releases/tag/v0.1.4
- The release has been archived on Zenodo: https://doi.org/10.5281/zenodo.8169295

I've double-checked everything. I think I've checked all the boxes on my end.
