### 2023-07-21 #5466 @samhforbes
Hi @rempsyc everything's looking really good.
Only comment I have on the editing of the paper is it would be great if you can edit the Quintana reference so it includes the DOI rather than just the OSF link, but otherwise all looks good.
