### 2023-07-21 2c307bd
add doi to paper [skip si]
