+++ paper/paper.bib
@@ -11,7 +11,7 @@
     title = {Five things about open and reproducible science that every early career researcher should know},
     author = {Quintana, D. S.},
     year = {2020},
-    url = {https://osf.io/2jt9u},
+    url = {https://osf.io/2jt9u; https://doi.org/10.17605/OSF.IO/DZTVQ},
   }
 
 @article{nuijten2016prevalence,
