+++ paper/paper.md
@@ -36,16 +36,16 @@ common in psychology research and social sciences more broadly.
 
 # Statement of need
 
-There are many reasons to use R ([R Core Team, 2022](#ref-base2021)) for
+There are many reasons to use R [@base2021] for
 analyzing and reporting data from research studies, such as being
-compatible with the ideals of open science ([Quintana,
-2020](#ref-quintana2020)). However, R has a major downside for novices:
+compatible with the ideals of open science [@quintana2020]. 
+However, R has a major downside for novices:
 its steep learning curve due to its programmatic interface, in contrast
 to perhaps more user-friendly point-and-click software. Of course, this
 flexibility is also a strength, as the R community can and does come
 together to produce packages that make using R increasingly easier and
-more user-friendly (e.g., the *easystats* ecosystem [Lüdecke et al.,
-2019/2023](#ref-easystatsPackage)). The {rempsyc} package (Really Easy
+more user-friendly [e.g., the _easystats_ 
+ecosystem @easystatsPackage]. The {rempsyc} package (Really Easy
 Methods for Psychology) contributes to this momentum by providing
 convenience functions that remove as much friction as possible between
 your script and your manuscript (in particular, if you are using
@@ -53,8 +53,8 @@ Microsoft Word).
 
 There are mainly three things that go into a manuscript: text, tables,
 and figures. {rempsyc} does not generate publication-ready text
-summarizing analyses; for this, see the {report} package ([Makowski et
-al., 2021/2023](#ref-reportPackage)). Instead, {rempsyc} focuses on the
+summarizing analyses; for this, see the {report} package [@reportPackage]. 
+Instead, {rempsyc} focuses on the
 production of publication-ready tables and figures. Below, I go over a
 few quick examples of those.
 
@@ -67,18 +67,17 @@ to their manuscript, or retype them manually. Yet, this approach
 increases the risks of copy-paste and retyping errors so common in
 psychology. This problem is not trivial given that according to some
 estimates, up to 50% of articles in psychology have at least one
-statistical error ([Nuijten et al., 2016](#ref-nuijten2016prevalence)).
+statistical error [@nuijten2016prevalence].
 Ideally, one should be able to format the table directly in R, and to
 export it to Word directly.
 
 Formatting a table properly in R is already a tedious and time-consuming
-task, but fortunately several packages take care of this step (e.g., the
-{broom} or {report} packages, [Makowski et al.,
-2021/2023](#ref-reportPackage); [Robinson et al., 2022](#ref-broom2022),
+task, but fortunately several packages take care of this step [e.g., the {broom} or 
+{report} packages, @reportPackage; @broom2022, and there are several others],
 and there are several others). Exporting these formatted tables to
 Microsoft Word remains a challenge however. Some packages do export to
-Word (e.g., the {apaTables} package [Stanley & Spence,
-2018](#ref-stanley2018reproducible)), but their formatting is often
+Word [e.g., the {apaTables} package 
+@stanley2018reproducible], but their formatting is often
 rigid especially when using analyzes or table formats that are not
 supported by default.
 
@@ -116,7 +115,10 @@ We can do the same with a {report} table.
 The {report} package provides quite comprehensive tables, so one may
 request an abbreviated table with the `'short'` argument. For
 convenience, it is also possible to highlight significant results for
-better visual discrimination, using the `'highlight'` argument.[1] Once
+better visual discrimination, using the `'highlight'` argument.^[This argument 
+can be used logically, as `'TRUE'` 
[…]
