+++ paper/paper.bib
@@ -26,7 +26,7 @@
 }
 
 @software{easystatsPackage,
-  title = {{{easystats}}: Streamline Model Interpretation, Visualization, and Reporting},
+  title = {{{easystats}}: An R Framework for Easy Statistical Modeling, Visualization, and Reporting},
   shorttitle = {{{easystats}}},
   author = {L\"udecke, Daniel and Makowski, Dominique and Ben-Shachar, Mattan S. and Patil, Indrajeet and Wiernik, Brenton M. and Bacher, Etienne and Thériault, Rémi },
   date = {2023-02-04T22:06:06Z},
