### 2023-07-24 2e20829
fix capital and doi url
