+++ paper/paper.bib
@@ -11,7 +11,7 @@
     title = {Five things about open and reproducible science that every early career researcher should know},
     author = {Quintana, D. S.},
     year = {2020},
-    url = {https://osf.io/2jt9u; https://doi.org/10.17605/OSF.IO/DZTVQ},
+    url = {https://doi.org/10.17605/OSF.IO/DZTVQ},
   }
 
 @article{nuijten2016prevalence,
@@ -26,7 +26,7 @@
 }
 
 @software{easystatsPackage,
-  title = {{{easystats}}: An R Framework for Easy Statistical Modeling, Visualization, and Reporting},
+  title = {{{easystats}}: An {R} Framework for Easy Statistical Modeling, Visualization, and Reporting},
   shorttitle = {{{easystats}}},
   author = {L\"udecke, Daniel and Makowski, Dominique and Ben-Shachar, Mattan S. and Patil, Indrajeet and Wiernik, Brenton M. and Bacher, Etienne and Thériault, Rémi },
   date = {2023-02-04T22:06:06Z},
