### 2023-07-24 #5466 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
