### 2023-07-25 #5466 @rempsyc
Thank you @samhforbes, @avakiai, @kinleyid, and @brunomontezano for making this possible! I had a great experience with the review process at JOSS and I would recommend it to my peers. Wishing you all the best
